Updates Personnes & Creating .JARs in doc "out"
