+++ .idea/libraries/Maven__org_postgresql_postgresql_42_3_1.xml
@@ -1,0 +1,14 @@
+<component name="libraryTable">
+  <library name="Maven: org.postgresql:postgresql:42.3.1" type="java-imported" external-system-id="Maven">
+    <properties groupId="org.postgresql" artifactId="postgresql" version="42.3.1" />
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/42.3.1/postgresql-42.3.1.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/42.3.1/postgresql-42.3.1-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/42.3.1/postgresql-42.3.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
