+++ Livre/src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Main-Class: fr.dauphine.miageif.MSA.Livre.LivreApplication
+
